Modularize cmake files; allow individual selection of apps to build
This commit does 2 major things. First, it extracts regions of cmake from the CMakeLists file into .cmake files that are included in-place from where they were extracted. This makes each file a bit easier to read and understand. Second, it changes the way app selection is done. Instead of always building every app, it provides an option to only build a single app. The syntax for building all apps is unchanged, and the tests have the same results as before. The purpose of this commit is to prepare the am-cakeml to be used in the seL4 environment and build system. This commit does not add any seL4-specific code, however.