Extract bash (and posix) dependent functionality
This creates a new structure, BashFunctions, subtracted from the old ManifestJsonConfig. The ManifestJsonConfig structure no longer depends on anything but am-cakeml source and json input, but it uses the BashFunctions structure. This is a first step towards adding configurability for an AM running in the seL4 environment, which does not provide a shell and is not posix-compliant.