Fix sneaky CryptoFFI bug
This bug amounts to a difference in input and output types for a function: DecryptOneShot. The reason for this bug is that this function is one of several that need to be exposed in an exact way. In other words, this function in Evercrypt crypto needs to match the samely-named function in OpenSSL crypto and any other type of crypto one might choose to use. However, I can't think of a way to ask the compiler to check that these files both fulfill the same interface. The result of this bug was an exotic type error way down the line that was essentially unreadable.